feat(components): use display-utils instead of individual declaration of const's | sas / bh | #2298
