--- packages/components/src/components/display/display-color.ts
@@ -1,2 +1,0 @@
-export const DISPLAY_COLORS = ['primary', 'inherit'] as const;
-export type DisplayColor = typeof DISPLAY_COLORS[number];
