--- packages/components/src/components/display/display-size.ts
@@ -1,2 +1,0 @@
-export const DISPLAY_SIZES = ['medium', 'large', 'inherit'] as const;
-export type DisplaySize = typeof DISPLAY_SIZES[number];
