--- packages/components/src/components/display/display-tag.ts
@@ -1,2 +1,0 @@
-export const DISPLAY_TAGS = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'] as const;
-export type DisplayTag = typeof DISPLAY_TAGS[number];
